requirements.txt
langchain-core
streamlit
dotenv
langchain-google-genai
langchain_community
langchain-ollama
langserve
uvicorn
fastapi
sse_starlette